no warning, and a bug with O3 optimize build also add d/d/d.hpp and d/d/d.cpp to test maketree  Changes to be committed: 	modified:   autocompile.cpp 	modified:   compile.O3.sh 	modified:   compile.bat 	modified:   compile.sh 	modified:   config.cpp 	modified:   config.hpp 	modified:   dirs.hpp 	modified:   file.hpp 	modified:   maketree.cpp 	modified:   process_output.hpp 	modified:   process_output_test.cpp 	modified:   test_project2/a/Makefile 	modified:   test_project2/b/.autocompile 	modified:   test_project2/b/Makefile 	modified:   test_project2/b/b.cpp 	modified:   test_project2/c/Makefile 	new file:   test_project2/d/d/Makefile 	new file:   test_project2/d/d/d.cpp 	new file:   test_project2/d/d/d.hpp 	modified:   test_project2/dirs

+++ dirs.hpp
@@ -19,7 +19,7 @@ private:
         vector<string> lines;
         if(read_all_lines(file, lines))
         {
-            for(int i = 0; i < lines.size(); i++)
+            for(size_t i = 0; i < lines.size(); i++)
             {
                 const string& s = lines[i];
                 if(!s.empty() && s[0] != '#')
